Created robust test to test ALL solvers (testing led development)
When I create a new solver, I can simply add it to the list of
solvers at the top of test_engine.py, then this test will run
on that solver too. For now, I have commented out the testing on
the hard tests, since backtracking takes too long to find these
solutions, hence it is not feasible to run this test prior to
every commit.