Created function to initialise candidates grid & wrote tests for it.
This function takes in a puzzle, and initialises the 'candidates grid',
where each item at a given index is the set of possible values for the
square at that index in the puzzle.

Explanation: We will use this candidates grid to perform candidate
elimination techniques, to aid with solving sudoku puzzles.