Renamed functions for readability
init_candidates_grid -> init_candidates
print_candidates_grid -> print_candidates
We chose to this naming since we use candidates as the
variable name for the candidates grid, hence we want the
function names to be consistent with the variable names.
We chose this variable name since we will be using this variable
lots, so I think candidates_grid is too long and is not required
for extra readability.