Wrote filler functions and tests
The changes in this commit are as follows:
- Wrote 'filler' function, which fills a puzzle using the candidate grid
(if there are empty squares which only have one candidate according to the candidate grid)
- Modified 'singles_filler' to use the filler function
- Wrote tests for filler function and singles_filler
- (actually just moved and slightly alterned the existing tests for singles_filler)

Explanation: This function will be useful when we have optimised
candidate grids which use an array of candidate elimination
techniques. Ie. We can optimise the candidate grid, then fill
in a puzzle using the grid.