Edited save_puzzle to make the directory in the filepath provided
(If it does not exist already). Also we changed the test for this
function to save files in the tests/test_puzzles/ directory, since
this directory is committed, allowing other people to run the tests.

Explanation for this change: We want solve_sudoku.py to save the files
in a directory solutions/ which will not be committed, hence the
save_puzzles function must make this directory first