Added functionality to find multiple solutions if desired and to save solutions
Modified tests/test_solver.py to accomodate and test this new functionality.
Specifically, we test that solve_sudoku can find multiple solutions, and
we modified the other tests to delete the files created by running the solver.

Potential issue: These tests would overwrite and delete files in
solutions/ if they are of the same names as the files used in the
tests. We will add a warning to README.md to specify this.